pyproject.toml
[project]
name = "casbin-tortoise-adapter"
version = "1.2.2"
description = "Tortoise ORM adapter for AsyncCasbin"
readme = "README.md"
keywords = [
    "casbin",
    "asyncio",
    "access control",
    "Tortoise ORM",
    "authorization",
    "pycasbin"
]
license = {text = "Apache-2.0"}
authors = [
    {name = "Elias Gabriel", email = "me@eliasfgabriel.com"},
]

requires-python = ">=3.7,<4.0"
dependencies = [
    "asynccasbin<2.0.0,>=1.1.2",
    "tortoise-orm[accel]>=0.19.1; python_version < \"3.8\"",
    "tortoise-orm[accel]>=0.20; python_version >= \"3.8\"",
]

[project.urls]
repository = "https://github.com/thearchitector/casbin-tortoise-adapter"
documentation = "https://github.com/thearchitector/casbin-tortoise-adapter#tortoise-orm-adapter-for-asynccasbin"
changelog = "https://github.com/thearchitector/casbin-tortoise-adapter/blob/main/CHANGELOG.md"

[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "pytest<8.0,>=7.1",
    "pytest-asyncio>=0.15.1",
    "asyncpg>=0.28.0",
]

[tool.pytest.ini_options]
addopts = "-ra -svv"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.pyright]
reportMissingTypeStubs = "none"
ignore = [ "tests" ]

[tool.mypy]
python_version = "3.7"
strict = true
disallow_subclassing_any = false

[[tool.mypy.overrides]]
module = "casbin.*"
ignore_missing_imports = true

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
